+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add data fetchers to `AVAILABLE_FOR_PLUGINS`
 - Update track list in `AVAILABLE_FOR_PLUGINS`
 - Correctly setup initial scales of vertical tracks when the width of a center track is zero.
+- If only vertical tracks are used, use the entire vertical space to display vertical tracks.
 
 _[Detailed changes since v1.11.5](https://github.com/higlass/higlass/compare/v1.11.7...develop)_